[Refactor] : 예외 구조 리팩토링
